Move `build_demo_graph` to `marker` module and remove unused imports from `dev_cmd`
- Relocated `build_demo_graph` function from `dev_cmd` to `plotting::marker` module.
- Removed unused imports from `dev_cmd` including `MarkerLevel` and several plotters dependencies.
- Updated color references in `build_demo_graph` to use `full_palette` instead of `colors`.
